change naming schema, rename count to band_count, don't allow multi band images as input for tile cache
