+++ app/tasks/raster_tile_cache_assets/utils.py
@@ -1,14 +1,18 @@
+import math
 import os
-from typing import List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
+import numpy as np
 from fastapi.logger import logger
 
 from app.crud.assets import create_asset
 from app.models.enum.assets import AssetType
+from app.models.enum.pixetl import DataType
 from app.models.pydantic.assets import AssetCreateIn
 from app.models.pydantic.creation_options import RasterTileSetSourceCreationOptions
 from app.models.pydantic.jobs import GDAL2TilesJob, Job
 from app.models.pydantic.metadata import RasterTileSetMetadata
+from app.models.pydantic.statistics import RasterStats
 from app.settings.globals import MAX_CORES, MAX_MEM, TILE_CACHE_BUCKET
 from app.tasks import Callback, callback_constructor
 from app.tasks.raster_tile_set_assets.utils import JOB_ENV, create_pixetl_job
@@ -182,3 +186,51 @@ def get_zoom_source_uri(
     return (
         [tile_uri_to_tiles_geojson(uri) for uri in source_uri] if source_uri else None
     )
+
+
+def generate_stats(stats) -> RasterStats:
+    if stats is None:
+        raise NotImplementedError()
+    return RasterStats(**stats)
+
+
+def convert_float_to_int(
+    stats: Optional[Dict[str, Any]],
+    source_asset_co: RasterTileSetSourceCreationOptions,
+) -> Tuple[RasterTileSetSourceCreationOptions, str]:
+    stats = generate_stats(stats)
+
+    assert len(stats.bands) == 1
+    stats_min = stats.bands[0].min
+    stats_max = stats.bands[0].max
+    value_range = math.fabs(stats_max - stats_min)
+
+    # Shift by 1 (and add 1 later) so any values of zero don't get counted as no_data
+    uint16_max = np.iinfo(np.uint16).max - 1
+    # Expand or squeeze to fit into a uint16
+    mult_factor = int(math.floor(uint16_max / value_range)) if value_range else 1
+
+    if isinstance(source_asset_co.no_data, list):
+        raise RuntimeError("Cannot apply colormap on multi band image")
+    elif source_asset_co.no_data is None:
+        old_no_data: str = "None"
+    elif source_asset_co.no_data == str(np.nan):
+        old_no_data = "np.nan"
+    else:
+        old_no_data = str(source_asset_co.no_data)
+
+    calc_str = (
+        f"(A != {old_no_data}).astype(bool) * "
+        f"(1 + (A - {stats_min}) * {mult_factor}).astype(np.uint16)"
+    )
+
+    source_asset_co.data_type = DataType.uint16
+    source_asset_co.no_data = 0
+
+    if source_asset_co.symbology and source_asset_co.symbology.colormap is not None:
+        source_asset_co.symbology.colormap = {
+            (1 + (float(k) - stats_min) * mult_factor): v
+            for k, v in source_asset_co.symbology.colormap.items()
+        }
+
+    return source_asset_co, calc_str
